Move test files to test SDK's protected methods
